:hammer: refactor: Changed create_post path to Scripts
